# [CIS-1830] Speed up bootstrap action