### 🔗 Issue Links

- https://stream-io.atlassian.net/browse/CIS-1830

### 🎯 Goal

To speed up bootstrap action

### 📝 Summary

- Change `mint` cache path
- Update `mint` cache key
- Move the `bootstrap` script to `Scripts` folder
- Move the `bootstrap` cache to the action file
- Add `set-build-image-var` action where possible

### 🎨 Showcase

Bootstrap time reduced from ≈4min to ≈40sec 🏎

| Before | After |
| ------ | ----- |
| ![before](https://user-images.githubusercontent.com/20794991/167385053-e0b2ce94-bab8-41de-900f-cc6261201ae3.png)|![after](https://user-images.githubusercontent.com/20794991/167386329-39497dc0-0013-4529-862e-ab45b52a5648.png)|

### ☑️ Contributor Checklist
- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [ ] Changelog is updated with client-facing changes
- [ ] New code is covered by unit tests
- [x] This change follows zero ⚠️ policy (required)
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

### 🎁 Meme

![gif](https://media.giphy.com/media/129s9ZDHNVjqmI/giphy.gif)